Begin world expansion
With noise semi-functional, we now need to start thinking about how to build
out the world data structures. This commit begins the work to transition away
from a static camera and world. It implements a basic square tile which holds
a vector of the "blocks" in that tile, and then builds out enough tiles to
fill the defined space for the world, initially 8192 * 2048.

The camera doesn't do anything more yet but its position has been defined so
this will likely be the next step.