Implement basic camera controls
Now that there is a world larger than the viewport, there needs to be a way to
navigate the world. This change implements basic keyboard and mouse control.

The keyboard control is far too slow to be useful and probably isn't necessary
but will be held on to for now.