Add threading for noise generation
Startup was significantly slowed by waiting for the noise to generate fresh
on every launch. Threading will be helpful in the future anyway, so I bit the
bullet and implemented the BS::thread-pool library, and am using it to wrap
the noise generation portion of start up.

Load time has gone from 10+ seconds to ~1.