Cleanup starting position, movement neuron work
Arbitrarily moved ants to start in an open spot on seed I'm testing with.
Starting area generation can come later, for now I'd like to focus on the ants.

Started the groundwork for world interaction. Added World::GetBlockAtPos to be
able to evaluate the environment. Added this to the movement neuron to prevent
movement out of the open starting area.