Implement tick thresholds
All ticks now accept an optional threshold variable. This will be used to
indicate how "strong" the signal is from the sending neuron.

The timer neuron will now send a signal every tick with a strength which is
related to how close the current timer is to the configured maximum delay.

It may be worth considering a feedback loop where the receiving neuron's tick
function returns a bool to indicate whether or not it did anything, allowing
the notifying neuron to change it's behavior with that information.