Timer Neuron fixes, task switching fixes
The method that was being used for checking time passed by the timer neuron
was failing after a seemingly random period of time. I'm not sure it's actually
fixed but I've made some adjustments and it seems to be behaving better.

There were also some issues with faulty logic overriding the task switching.
A few of these have been resolved although at this point the ant will not start
a second search for food.