Fix going idle after delivering food
Added the searching task to the wander movement so that ants will have a chance
to get in range of what they're looking for. This behavior shines a light on
why pheromone tracking is needed.

For now, manually boosted search distance to a much longer range. The search
algorithm will probably need attention before the ranges get very long.