Large refactor... again...
Things have mostly recovered from the rabbit hole I fell down this weekend. The
tests are starting to get implemented and work, I've split up some of the files
for readibility but more work is required there. Definitely more work to come
in this regard, but I'd like to make some actual progress on the sim itself.