More structure cleanup, fixed food placement
Made file naming and class naming more consistent. Split the movement neuron
into its own file. More cleanup like this will continue.

Focused on ensuring neurons only include code for tasks they specifically
handle. Task neuron handles the state machine, movement neuron deals with
moving only, and so forth. This helped in pinning down the food placement issue.

Food placement was failing because the destination and placement location were
set to the same spot. The ant now navigates to one block away from the
placement destination.

This brings the whole application back to the state it was before the major
rework and cleanup. The ant will deliver multiple pieces of food at which point
the navigation just fails. So, the cleanup hasn't helped that issue which is
annoying, but at the very least it's interesting to see it back in the exact
same state. Time to finally knock this one out.