SOLID refactoring, laying async groundwork
Although the current implementation is currently functional, the program is
single threaded and blocks during long operations like A* pathfinding. While
it is functional, it also needs a fair amount of extension shortly and the
codebase is not well designed or prepared for that.

This commit is the first to move towards better design using SOLID principles,
which will help with both the extension/addition of more functionality per
module as well as the implementation of asynchronus tasks.